remove !important font-size rule #70
@ofZach this rule is causing lots of font size-related strangeness
throughout the site, so I'd like to just remove it. Is the intent to
make the type generally larger? If so I can do that via the bootstrap
config, which will scale everything else proportionally. Right now this
rule squashes down everything in a <p> tag kind of inconsistently, and
also has a chaotic recursive effect because of the `em` units.