fix 7 broken urls in /people/#filter=.summer-2016
`http://` was left out of 7 different links to personal websites. So if
a url was listed say as `jonl.org`, it would redirect to
`http://sfpc.io/people/jonl.org` (which does not exist) instead of
`http://jonl.org`. I've added the http:// back into those 7 different
urls which affects the personal websites of Jonathan, Krista, Matt,
Max, Meina, Melanie, and Oren.

I think this is a recurring problem on the website (not under /people's
webpages, but elsewhere).